Support to get usd-vnd exchange rate
