+++ src/main/java/com/techlooper/service/CurrencyService.java
@@ -1,0 +1,9 @@
+package com.techlooper.service;
+
+/**
+ * Created by phuonghqh on 5/28/15.
+ */
+public interface CurrencyService {
+
+  Long usdToVndRate();
+}
