+++ src/main/java/com/techlooper/service/impl/PromotionServiceImpl.java
@@ -1,9 +1,10 @@
 package com.techlooper.service.impl;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.techlooper.entity.SalaryReview;
 import com.techlooper.model.CitibankCreditCardPromotion;
+import com.techlooper.model.PaymentMethod;
 import com.techlooper.repository.elasticsearch.SalaryReviewRepository;
+import com.techlooper.service.CurrencyService;
 import com.techlooper.service.PromotionService;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
@@ -11,7 +12,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.MailSender;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
 
 import javax.annotation.Resource;
 import java.io.IOException;
@@ -46,30 +46,26 @@ public class PromotionServiceImpl implements PromotionService {
   private Map<Long, String> locationMap;
 
   @Resource
-  private Long usdToVndRate;
+  private CurrencyService currencyService;
 
   public boolean placePromotion(CitibankCreditCardPromotion citibankCreditCardPromotion) throws IOException, TemplateException {
     SalaryReview salaryReview = salaryReviewRepository.findOne(citibankCreditCardPromotion.getSalaryReviewId());
-    Long netIncome = usdToVndRate * salaryReview.getNetSalary();
-    if (citibankCreditCardPromotion.getAgree() != Boolean.TRUE
-      || netIncome < minIncome
-      || !acceptCityIds.contains(salaryReview.getLocationId().toString())) {
+    Long netIncome = currencyService.usdToVndRate() * salaryReview.getNetSalary();
+    if (citibankCreditCardPromotion.getAgree() != Boolean.TRUE ||
+      citibankCreditCardPromotion.getPaymentMethod() != PaymentMethod.BANK_TRANSFER ||
+      netIncome < minIncome ||
+      !acceptCityIds.contains(salaryReview.getLocationId().toString())) {
       return false;
     }
 
     citibankCreditCardPromotion.setNetIncome("VND " + netIncome);
-    citibankCreditCardPromotion.setLocation(locationMap.get(salaryReview.getLocationId()));
+
+    String location = locationMap.get(salaryReview.getLocationId());
+    citibankCreditCardPromotion.setLocation(location);
     StringWriter stringWriter = new StringWriter();
     citibankCreditCardPromotionTemplate.process(citibankCreditCardPromotion, stringWriter);
     citibankCreditCardPromotionMailMessage.setText(stringWriter.toString());
     mailSender.send(citibankCreditCardPromotionMailMessage);
     return true;
   }
-
-//  public static void main(String[] args) {
-//    JsonNode abc = new RestTemplate().getForEntity("http://www.freecurrencyconverterapi.com/api/v3/convert?q=USD_VND", JsonNode.class).getBody();
-//    System.out.println(abc.toString());
-//    System.out.println(abc.findPath("val").asDouble());
-//    System.out.println(abc.path("/results/USD_VND/val").asText());
-//  }
 }
