+++ src/main/resources/dev/techlooper.properties
@@ -22,7 +22,7 @@ webapp.resource.location = /public/
 
 core.textEncryptor.password = XR3hBZnJ6gLH
 
-mail.citibank.cc_promotion.to = phuonghqh@gmail.com
+mail.citibank.cc_promotion.to = thu.hoang@navigosgroup.com
 mail.citibank.cc_promotion.subject = VNW App Salary
 
 mail.form = techlooperbyvnw@gmail.com
