+++ src/main/resources/production/techlooper.properties
@@ -24,7 +24,7 @@ core.textEncryptor.password = XR3hBZnJ6gLH
 
 salaryEvaluation.allowToSave = true
 
-mail.citibank.cc_promotion.to = phuonghqh@gmail.com
+mail.citibank.cc_promotion.to = thu.hoang@navigosgroup.com
 mail.citibank.cc_promotion.subject = VNW App Salary
 
 mail.form = techlooperbyvnw@gmail.com
